Converts types package to module Renames package to _types, adds types module, removes redundant init, and resolves imports. Pretty much everything touches the types package, so this one's a doozy. Unclear to me if there was a better way to break this up.
