+++ src/pyrite/_rendering/view_plane.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 
-from ..types.view_bounds import CameraViewBounds
+from pyrite._types.view_bounds import CameraViewBounds
 
 if TYPE_CHECKING:
     from pygame import Rect
-    from ..types.bounds import CullingBounds
+    from pyrite._types.bounds import CullingBounds
 
 
 class ViewPlane(CameraViewBounds):
