+++ src/pyrite/_transform/transform.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 
 from pygame import Vector2
 
-from ..types import transform
+from pyrite._types.transform import TransformLike
 
 if TYPE_CHECKING:
     from pygame.typing import Point
@@ -53,7 +53,7 @@ class Transform:
         self._scale = Vector2(new_scale)
 
     def __eq__(self, value: object) -> bool:
-        if not isinstance(value, transform.TransformLike):
+        if not isinstance(value, TransformLike):
             return False
         return (
             value.position == self._position
@@ -65,9 +65,7 @@ class Transform:
         return Transform(self._position, self._rotation, self._scale)
 
     @staticmethod
-    def generalize(
-        branch: transform.TransformLike, root: transform.TransformLike
-    ) -> Transform:
+    def generalize(branch: TransformLike, root: TransformLike) -> Transform:
         """
         Applies a root transform to a local transform, converting it into the same
         relative space.
@@ -126,16 +124,14 @@ class Transform:
         rotated_position = scaled_position.rotate(self.rotation)
         return self.position + rotated_position
 
-    def __mul__(self, other_transform: transform.TransformLike) -> Transform:
+    def __mul__(self, other_transform: TransformLike) -> Transform:
         return Transform.generalize(other_transform, self)
 
-    def __rmul__(self, other_transform: transform.TransformLike) -> Transform:
+    def __rmul__(self, other_transform: TransformLike) -> Transform:
         return Transform.generalize(self, other_transform)
 
     @staticmethod
-    def localize(
-        branch: transform.TransformLike, root: transform.TransformLike
-    ) -> Transform:
+    def localize(branch: TransformLike, root: TransformLike) -> Transform:
         """
         Given two Transforms in the same relative space, finds the Transform local to
         the root that is equivalent of this Transform.
@@ -189,10 +185,10 @@ class Transform:
         rotated_position = translated_position.rotate(self.rotation)
         return rotated_position / self.scale.elementwise()
 
-    def __truediv__(self, other: transform.TransformLike) -> Transform:
+    def __truediv__(self, other: TransformLike) -> Transform:
         return Transform.localize(self, other)
 
-    def __rtruediv__(self, other: transform.TransformLike) -> Transform:
+    def __rtruediv__(self, other: TransformLike) -> Transform:
         return Transform.localize(other, self)
 
     def __repr__(self) -> str:
