+++ src/pyrite/_types/bounds.py
