+++ src/pyrite/_types/component.py
