+++ src/pyrite/_types/entity.py
