+++ src/pyrite/_types/renderer.py
