+++ src/pyrite/_types/shape.py
