+++ src/pyrite/_types/system.py
