+++ src/pyrite/_types/view_bounds.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ..types.bounds import CullingBounds
+    from pyrite._types.bounds import CullingBounds
 
 
 class CameraViewBounds(ABC):
